app/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nmea-handler"
version = "0.1.0"
description = "A BlueOS extension for handling NMEA messages from serial devices"
authors = [
    {name = "Tony White", email = "tony.white@bluerobotics.com"}
]
dependencies = [
    "litestar>=0.27.0",
    "pyserial>=3.5",
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["app"]

